feat: API 스펙 변경 대응

+++ app/src/main/java/com/choius323/saisai/data/course/remote/model/PopularChallengeItemDto.kt
@@ -9,7 +9,7 @@ data class PopularChallengeItemDto(
     val courseName: String,
     val distance: Double,
     val endedAt: String,
-    val estimatedTime: Int,
+    val estimatedTime: Double,
     val level: Int,
     val challengerCount: Int,
     val sigun: String,
